--- app/src/main/java/com/shumikhin/rainandsun/ui/main/MainViewModel.kt
@@ -1,7 +1,0 @@
-package com.shumikhin.rainandsun.ui.main
-
-import androidx.lifecycle.ViewModel
-
-class MainViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
