+++ app/src/main/res/layout/main_fragment.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.MainFragment">
+    tools:context=".view.MainFragment">
 
     <TextView
         android:id="@+id/message"
